cloud_function/requirements.txt
pandas
scikit-learn
joblib
google-cloud-storage
